+++ lib/src/systems/stream_x.dart
@@ -1,7 +1,6 @@
-import 'dart:async';
-
-import 'system.dart';
-import '../types/types.dart';
+import 'dart:async' show Stream, StreamController, StreamSubscription;
+import 'system.dart' show System;
+import '../types/types.dart' show Moment;
 
 extension StreamX<State, Event> on System<State, Event> {